+++ src/main/java/com/example/applefarm_/admin/controller/AdminController.java
@@ -3,6 +3,7 @@ package com.example.applefarm_.admin.controller;
 import com.example.applefarm_.admin.service.AdminServiceImpl;
 import com.example.applefarm_.user.dto.UserResponseDto;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,6 +12,7 @@ import java.util.List;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/admin")
+@PreAuthorize("hasRole('ROLE_ADMIN')")
 public class AdminController {
     private final AdminServiceImpl adminrService;
 
@@ -23,15 +25,5 @@ public class AdminController {
     public List<UserResponseDto> findBySellerList() {
         return adminrService.findSellerList();
     }
-
-    @PutMapping("/modifideroleCustomer/{id}")
-    public void modifideroleCustomer(@PathVariable Long id){
-        adminrService.modifideroleCustomer(id);
-    }
-    @PutMapping("/modifideroleSeller/{id}")
-    public void modifideroleSeller(@PathVariable Long id){
-        adminrService.modifideroleSeller(id);
-    }
-
 
 }
