+++ src/main/java/com/example/applefarm_/admin/service/AdminService.java
@@ -12,6 +12,4 @@ public interface AdminService {
     @Transactional
     List<UserResponseDto> findSellerList();
 
-    void modifideroleCustomer(Long id);
-    void modifideroleSeller(Long id);
 }
