+++ src/main/java/com/example/applefarm_/user/service/UserServiceImpl.java
@@ -30,6 +30,9 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static com.example.applefarm_.user.entitiy.UserRoleEnum.CUSTOMER;
+import static com.example.applefarm_.user.entitiy.UserRoleEnum.SELLER;
+
 @Service
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
@@ -105,5 +108,26 @@ public class UserServiceImpl implements UserService {
         user.update(userProfileRequestDto);
         userRepository.save(user);
         return new UserProfileResponseDto(user);
+    }
+
+    @Override
+    public void modifideroleCustomer(Long id) throws IllegalArgumentException {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("등록된 사용자가 없습니다.")
+        );
+        if(user.getRole() == CUSTOMER){
+            user.changeSellerByCustomer();
+        }else {
+            throw new IllegalArgumentException("현재 사용자는 Customer가 아닙니다.");
+        }
+    }
+    @Override
+    public void modifideroleSeller(Long id) throws IllegalArgumentException {
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("등록된 사용자가 없습니다.")
+        );
+        if(user.getRole() == SELLER){
+            user.changeCustomerBySeller();
+        }else {throw new IllegalArgumentException("현재 사용자는 Seller가 아닙니다.");}
     }
 }
